requirements.txt
dotenv
musicbrainzngs
mutagen
openai
pyacoustid
requests